scaladoc, move logic to different files
